+++ src/main/java/roomescape/common/exception/AlreadyExistException.java
@@ -1,5 +1,7 @@
 package roomescape.common.exception;
 
+import roomescape.common.exception.error.GeneralErrorCode;
+
 public class AlreadyExistException extends CustomException {
 
     public AlreadyExistException(String message) {
