+++ src/main/java/roomescape/common/exception/AuthenticationException.java
@@ -1,5 +1,7 @@
 package roomescape.common.exception;
 
+import roomescape.common.exception.error.GeneralErrorCode;
+
 public class AuthenticationException extends CustomException {
 
     public AuthenticationException(String message) {
