+++ src/main/java/roomescape/common/exception/CustomException.java
@@ -1,6 +1,7 @@
 package roomescape.common.exception;
 
 import lombok.Getter;
+import roomescape.common.exception.error.ErrorCode;
 
 @Getter
 public abstract class CustomException extends RuntimeException {
