+++ src/main/java/roomescape/common/exception/InvalidInputException.java
@@ -1,5 +1,7 @@
 package roomescape.common.exception;
 
+import roomescape.common.exception.error.GeneralErrorCode;
+
 public class InvalidInputException extends CustomException {
 
     public InvalidInputException(String message) {
