+++ src/main/java/roomescape/common/exception/PaymentException.java
@@ -1,5 +1,8 @@
 package roomescape.common.exception;
 
+import roomescape.common.exception.error.ErrorCode;
+import roomescape.common.exception.error.PaymentErrorCode;
+
 public class PaymentException extends CustomException {
 
     public PaymentException(String message, ErrorCode errorCode) {
