+++ src/main/java/roomescape/common/exception/TimeOutException.java
@@ -1,5 +1,7 @@
 package roomescape.common.exception;
 
+import roomescape.common.exception.error.ErrorCode;
+
 public class TimeOutException extends ConnectionException {
 
     public TimeOutException(String message) {
