+++ src/test/java/roomescape/payment/service/client/PaymentServerTest.java
@@ -7,14 +7,12 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
 import java.time.Duration;
-import java.util.concurrent.TimeUnit;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
 import okhttp3.mockwebserver.SocketPolicy;
 import org.apache.http.HttpHeaders;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -23,7 +21,7 @@ import org.springframework.http.MediaType;
 import org.springframework.http.client.SimpleClientHttpRequestFactory;
 import org.springframework.web.client.RestClient;
 import roomescape.common.exception.ConnectionException;
-import roomescape.common.exception.PaymentErrorCode;
+import roomescape.common.exception.error.PaymentErrorCode;
 import roomescape.common.exception.PaymentException;
 import roomescape.payment.service.dto.PaymentClientErrorResponse;
 import roomescape.payment.service.dto.PaymentClientResponse;
